Se separan options de topnav en componentes.
