+++ src/layouts/Sidebar.vue
@@ -1,8 +1,3 @@
-<script setup lang="ts">
-    import { RouterLink } from 'vue-router'
-    import MenuItem from '../components/MenuItem.vue';
-</script>
-
 <template>
     <div class="flex flex-col h-screen">
         <div id="UserOptions" class="pl-6 w-[240px] bg-black">
@@ -18,18 +13,23 @@
                 <RouterLink to="/search">
                     <MenuItem class="ml-[1px]" :iconSize="24" name="Buscar" iconString="search" pageUrl="/search" />
                 </RouterLink>
-                <RouterLink to="/library">
+                <RouterLink to="/library" v-if="mainStore.$state.token">
                     <MenuItem class="ml-[1px]" :iconSize="24" name="Mis canciones" iconString="library" pageUrl="/library" />
                 </RouterLink>
                 
                 <div class="border-b border-b-gray-700 w-[200px]"></div>
-                <div class="mt-4">
-                    <MenuItem :iconSize="24" name="Crear Playlist" iconString="playlist" pageUrl="/playlist" />
-                    <MenuItem :iconSize="24" name="Subir Canción" iconString="upload" pageUrl="/addsong" class="ml-[1px]"/>
+                <h2 v-if="!mainStore.$state.token" class="text-white pt-5">¡Inicia sesión para más funcionalidades!</h2>
+                <div v-if="mainStore.$state.token">
+                    <div class="mt-4">
+                        <MenuItem :iconSize="24" name="Crear Playlist" iconString="playlist" pageUrl="/playlist" />
+                        <MenuItem :iconSize="24" name="Subir Canción" iconString="upload" pageUrl="/addsong" class="ml-[1px]"/>
+                    </div>
                 </div>
             </ul>
         </div>
-        <div id="UserPlaylists" class="h-[100%] pl-6 w-[240px] bg-black">
+
+        <!-- User logged -->
+        <div v-if="mainStore.$state.token" id="UserPlaylists" class="h-[100%] pl-6 w-[240px] bg-black">
             
             <div class="border-b border-b-gray-700 w-[200px]"></div>
                 <ul class="mt-6">
@@ -39,5 +39,18 @@
                     <li class="font-semibold text-[15px] mt-3 text-gray-300 hover:text-white">Tu Playlist #4</li>
                 </ul>
         </div>
+
+        <!-- Guest -->
+        <div v-if="!mainStore.$state.token" id="UserPlaylists" class="h-[100%] pl-6 w-[240px] bg-black text-white">
+            <span></span>
+        </div>
     </div>
 </template>
+
+<script setup lang="ts">
+    import { RouterLink } from 'vue-router'
+    import MenuItem from '../components/MenuItem.vue';
+    import { useMainStore } from '@/stores/main';
+
+    const mainStore = useMainStore();
+</script>
